Split ComponentInstaller.cs
