+++ tools/Installer/Installer/Actions/IDockerActions.cs
@@ -1,0 +1,12 @@
+﻿using Abs.CommonCore.Contracts.Json.Installer;
+
+namespace Abs.CommonCore.Installer.Actions;
+public interface IDockerActions
+{
+    bool WaitForDockerContainersHealthy { get; set; }
+
+    Task DockerSystemPruneAsync();
+    Task ExecuteDockerComposeAsync(string rootLocation, ComponentAction action);
+    Task RunDockerComposeCommandAsync(Component component, string rootLocation, ComponentAction action);
+    Task StopAllContainersAsync();
+}
